app_grpc/requirements.txt
# app_grpc/requirements.txt
grpcio>=1.40.0,<2.0.0
protobuf>=3.15.0,<4.0.0